feat: add remote mapping components for amazon product types
